split tree into smaller .rs files
