+++ src/middlewares/versioned_flat_key_value/pending_part/tree/mod.rs
@@ -1,16 +1,18 @@
+mod add_node_support;
+mod change_root_support;
+mod checkout_support;
+mod key_value_store_manager_support;
 mod node;
 
 pub type SlabIndex = usize;
 
-use std::collections::{BTreeMap, BTreeSet, HashMap, VecDeque};
+use std::collections::HashMap;
 
 use slab::Slab;
 
 use self::node::TreeNode;
 
-use super::pending_schema::{
-    ApplyMap, ApplyRecord, KeyValueMap, PendingKeyValueSchema, RecoverMap, Result as PendResult,
-};
+use super::pending_schema::{PendingKeyValueSchema, Result as PendResult};
 use super::PendingError;
 
 pub struct Tree<S: PendingKeyValueSchema> {
@@ -87,247 +89,5 @@ impl<S: PendingKeyValueSchema> Tree<S> {
         }
 
         slab_indices
-    }
-}
-
-// methods to support VersionedMap::add_node()
-impl<S: PendingKeyValueSchema> Tree<S> {
-    pub fn add_root(
-        &mut self,
-        commit_id: S::CommitId,
-        modifications: RecoverMap<S>,
-    ) -> PendResult<(), S> {
-        // return error if there is root
-        if self.has_root() {
-            return Err(PendingError::MultipleRootsNotAllowed);
-        }
-        // PendingError::CommitIdAlreadyExists(_) cannot happend because no root indicates no node
-
-        // new root
-        let root = TreeNode::new_root(commit_id, modifications, self.height_of_root);
-
-        // add root to tree
-        let slab_index = self.nodes.insert(root);
-        self.index_map.insert(commit_id, slab_index);
-
-        Ok(())
-    }
-
-    pub fn add_non_root_node(
-        &mut self,
-        commit_id: S::CommitId,
-        parent_commit_id: S::CommitId,
-        modifications: RecoverMap<S>,
-    ) -> PendResult<(), S> {
-        // return error if parent_commit_id does not exist
-        let parent_slab_index = self.get_slab_index_by_commit_id(parent_commit_id)?;
-
-        // return error if commit_id exists
-        if self.contains_commit_id(&commit_id) {
-            return Err(PendingError::CommitIdAlreadyExists(commit_id));
-        }
-
-        // new node
-        let node = TreeNode::new_non_root(
-            commit_id,
-            parent_slab_index,
-            self.get_node_by_slab_index(parent_slab_index).get_height() + 1,
-            modifications,
-        );
-
-        // add node to tree
-        let slab_index = self.nodes.insert(node);
-        self.index_map.insert(commit_id, slab_index);
-        self.nodes[parent_slab_index].insert_child(slab_index);
-
-        Ok(())
-    }
-}
-
-// methods to support VersionedMap::change_root()
-impl<S: PendingKeyValueSchema> Tree<S> {
-    // excluding target
-    fn find_path(&self, target_slab_index: SlabIndex) -> Vec<(S::CommitId, KeyValueMap<S>)> {
-        let mut target_node = self.get_node_by_slab_index(target_slab_index);
-        let mut path = VecDeque::new();
-        while let Some(parent_slab_index) = target_node.get_parent() {
-            target_node = self.get_node_by_slab_index(parent_slab_index);
-            path.push_front((target_node.get_commit_id(), target_node.get_updates()));
-        }
-        path.into()
-    }
-
-    // todo: test
-    #[allow(clippy::type_complexity)]
-    pub fn change_root(
-        &mut self,
-        commit_id: S::CommitId,
-    ) -> PendResult<(usize, Vec<(S::CommitId, KeyValueMap<S>)>), S> {
-        let slab_index = self.get_slab_index_by_commit_id(commit_id)?;
-
-        // root..=new_root's parent
-        let to_commit = self.find_path(slab_index);
-
-        // early return if new_root == root
-        if to_commit.is_empty() {
-            return Ok((
-                self.nodes[slab_index].get_height() - to_commit.len(),
-                to_commit,
-            ));
-        }
-
-        // subtree of new_root
-        let to_maintain_vec = self.bfs_subtree(slab_index);
-        let to_maintain = BTreeSet::from_iter(to_maintain_vec);
-
-        // remove: tree - subtree of new_root
-        let mut to_remove = Vec::new();
-        for (idx, _) in self.nodes.iter() {
-            if !to_maintain.contains(&idx) {
-                to_remove.push(idx);
-            }
-        }
-        for idx in to_remove {
-            self.index_map
-                .remove(&self.nodes.remove(idx).get_commit_id());
-        }
-
-        // set new_root's parent as None
-        self.nodes[slab_index].set_parent(None);
-
-        // (height of new_root's parent, root..=new_root's parent)
-        Ok((
-            self.nodes[slab_index].get_height() - to_commit.len(),
-            to_commit,
-        ))
-    }
-}
-
-// Internal Tree methods supporting
-// helper methods in VersionedMap for
-// implementing `KeyValueStoreManager` for `VersionedStore`.
-impl<S: PendingKeyValueSchema> Tree<S> {
-    pub fn iter_historical_changes<'a>(
-        &'a self,
-        commit_id: &S::CommitId,
-        key: &'a S::Key,
-    ) -> PendResult<impl 'a + Iterator<Item = (S::CommitId, &S::Key, Option<S::Value>)>, S> {
-        let mut node_option = Some(self.get_node_by_commit_id(*commit_id)?);
-        let mut path = Vec::new();
-        while let Some(node) = node_option {
-            if let Some(value) = node.get_modified_value(key) {
-                path.push((node.get_commit_id(), key, value));
-            }
-            node_option = self.get_parent_node(node);
-        }
-        Ok(path.into_iter())
-    }
-
-    pub fn get_versioned_key(
-        &self,
-        commit_id: &S::CommitId,
-        key: &S::Key,
-    ) -> PendResult<Option<Option<S::Value>>, S> {
-        let mut node_option = Some(self.get_node_by_commit_id(*commit_id)?);
-        while let Some(node) = node_option {
-            if let Some(value) = node.get_modified_value(key) {
-                return Ok(Some(value));
-            }
-            node_option = self.get_parent_node(node);
-        }
-        Ok(None)
-    }
-
-    pub fn discard(&mut self, commit_id: S::CommitId) -> PendResult<(), S> {
-        let slab_index = self.get_slab_index_by_commit_id(commit_id)?;
-        let node = self.get_node_by_slab_index(slab_index);
-        if let Some(parent_of_discard) = node.get_parent() {
-            let to_remove = self.bfs_subtree(slab_index);
-            for idx in to_remove {
-                self.index_map
-                    .remove(&self.nodes.remove(idx).get_commit_id());
-            }
-            let parent_node = self.get_mut_node_by_slab_index(parent_of_discard);
-            parent_node.remove_child(&slab_index);
-            Ok(())
-        } else {
-            Err(PendingError::RootShouldNotBeDiscarded)
-        }
-    }
-}
-
-// methods to support VersionedMap::checkout_current()
-impl<S: PendingKeyValueSchema> Tree<S> {
-    pub fn get_apply_map_from_root_included(
-        &self,
-        target_commit_id: S::CommitId,
-    ) -> PendResult<ApplyMap<S>, S> {
-        let mut target_node = self.get_node_by_commit_id(target_commit_id)?;
-        let mut commits_rev = BTreeMap::new();
-        target_node.export_commit_data(&mut commits_rev);
-        while let Some(parent_slab_index) = target_node.get_parent() {
-            target_node = self.get_node_by_slab_index(parent_slab_index);
-            target_node.export_commit_data(&mut commits_rev);
-        }
-        Ok(commits_rev)
-    }
-
-    // correctness based on single root
-    #[allow(clippy::type_complexity)]
-    pub(super) fn collect_rollback_and_apply_ops(
-        &self,
-        current_commit_id: S::CommitId,
-        target_commit_id: S::CommitId,
-    ) -> PendResult<(BTreeMap<S::Key, Option<ApplyRecord<S>>>, ApplyMap<S>), S> {
-        let mut current_node = self.get_node_by_commit_id(current_commit_id).unwrap();
-        let mut target_node = self.get_node_by_commit_id(target_commit_id)?;
-        let mut rollbacks = BTreeMap::new();
-        let mut commits_rev = BTreeMap::new();
-        while current_node.get_height() > target_node.get_height() {
-            current_node.export_rollback_data(&mut rollbacks);
-            current_node = self.get_parent_node(current_node).unwrap();
-        }
-        while target_node.get_height() > current_node.get_height() {
-            target_node.export_commit_data(&mut commits_rev);
-            target_node = self.get_parent_node(target_node).unwrap();
-        }
-        while current_node.get_commit_id() != target_node.get_commit_id() {
-            current_node.export_rollback_data(&mut rollbacks);
-            current_node = self.get_parent_node(current_node).unwrap();
-            target_node.export_commit_data(&mut commits_rev);
-            target_node = self.get_parent_node(target_node).unwrap();
-        }
-        // check rollbacks' old_commit_id because TreeNodes are deleted
-        // in a lazy way with respect to TreeNodes.modifications
-        // todo: test this lazy method
-        for (_, old_commit_id_option) in rollbacks.iter_mut() {
-            if let Some(ref old_commit_id) = old_commit_id_option {
-                if !self.contains_commit_id(old_commit_id) {
-                    *old_commit_id_option = None;
-                }
-            }
-        }
-
-        let rollbacks_with_value: BTreeMap<_, _> = rollbacks
-            .into_iter()
-            .map(|(k, old_cid_opt)| match old_cid_opt {
-                None => (k, None),
-                Some(rollback_cid) => {
-                    let rollback_value = self.get_by_commit_id(rollback_cid, &k).unwrap().unwrap();
-                    (
-                        k,
-                        Some(ApplyRecord::<S> {
-                            value: rollback_value,
-                            commit_id: rollback_cid,
-                        }),
-                    )
-                }
-            })
-            .collect();
-
-        // rollbacks or commits_rev may be empty,
-        // they contain current and target (if they are not lca), respectively,
-        // but they do not contain lca
-        Ok((rollbacks_with_value, commits_rev))
     }
 }
